Refactor Activity-ActivityType Relationship
Updated the Prisma schema to reflect a shift from using an
auto-incrementing integer 'id' to identifying ActivityType by unique
'name' field. Consequently, changed the relation in the Activity model
to link with 'activityTypeName' instead of the 'activityTypeId'. This
modification simplifies the relationship between Activities and
ActivityTypes, providing more meaningful and human-readable
associations. Also introduced a 'label' field to ActivityType for
additional descriptive purposes.